+++ doorstop/gui/test/test_all.py
@@ -7,7 +7,6 @@ Integration tests for the doorstop.cli package.
 import unittest
 from unittest.mock import patch, Mock
 
-import os
 import sys
 import imp
 
@@ -41,7 +40,7 @@ class TestImport(unittest.TestCase):  # pylint: disable=R0904
         """Verify tkinter import errors are handled."""
         sys.modules['tkinter'] = Mock(side_effect=ImportError)
         imp.reload(gui)
-        self.assertFalse(gui._run(None, None, None))  # pylint: disable=W0212
+        self.assertFalse(gui._run(None, None, lambda x: False))  # pylint: disable=W0212
         self.assertIsInstance(gui.tk, Mock)